requirements.txt
# Django and REST Framework
Django==5.2.5
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.10

# Caching
redis==6.3.0

# Background tasks
celery==5.5.3

# Configuration management
python-decouple==3.8

# HTTP requests
requests==2.32.4

# API documentation
drf-yasg==1.21.10

# Additional dependencies
PyYAML==6.0.2
uritemplate==4.2.0
inflection==0.5.1
packaging==25.0
pytz==2025.2

# Development dependencies (optional)
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3